chore: keep experimental tests local
